requirements.txt
openai>=1.50.0
python-pptx>=1.0.0
pydantic>=2.0.0
pytest>=8.0.0
python-dotenv>=1.0.0
